// fe_inventarios_gei/src/app/models/factorEmision.ts
export class factorEmision {
  nIdFactorEmision: Number;
  oFuenteEmision: any = {};
  sTipoFactor: String;
  sDescripcion: String;
  bdFeCO2: Number;
  bdFeCH4: Number;
  bdFeN2O: Number;
  bdValorConversion: Number;
  bdBiocombustible: Number;
  sUnidadFe: String;
  sUnidadVcn: String;
  sFechaMod: String;
  busqueda: any;
}